enhance dryrun

+++ scripts/deploy.sh
@@ -2,6 +2,12 @@
 set -xeuo pipefail
 trap 'echo "命令失败: 行 $LINENO"; exit 1' ERR
 
+DRYRUN=${DRYRUN:-false}
+if [ "$DRYRUN" == "true" ]; then
+  echo "DRYRUN 模式: $DRYRUN"
+  echo "DRYRUN 模式下，不执行实际部署，只打印部署命令和检查参数是否正确"
+fi
+
 # 确保可执行存在
 command -v polycli >/dev/null 2>&1 || { echo "未找到 polycli"; exit 1; }
 command -v jq >/dev/null 2>&1 || { echo "未找到 jq"; exit 1; }
@@ -80,19 +86,26 @@ export DEPLOY_PARAMETERS_SALT=0x$(openssl rand -hex 32)
 # sed "s/{{PRIVATE_KEY}}/$PRIVATE_KEY/g ; s/{{L2_CHAIN_ID}}/$L2_CHAIN_ID/g ; s/{{DEPLOY_PARAMETERS_SALT}}/$DEPLOY_PARAMETERS_SALT/g" params.template.yml > "$TEMP_CONFIG"
 # echo $L2_CONFIG >> $TEMP_CONFIG
 
-envsubst <params.template.yml >$TEMP_CONFIG
-
-echo "generated $TEMP_CONFIG"
-
 # 运行 kurtosis
-# kurtosis run --cli-log-level debug -v EXECUTABLE --enclave op-eth github.com/wangdayong228/optimism-package@8d97b22f5bce73106fea4d3cc063486cca359928 --args-file "$TEMP_CONFIG" 2>&1 > "$LOG_FILE"
-kurtosis run --cli-log-level debug -v EXECUTABLE --enclave $1 --args-file $TEMP_CONFIG github.com/Pana/kurtosis-cdk@aa5f6f39dd8fa6157abe5736d81a2c9eda1536fc 2>&1 >$LOG_FILE
-
-# 设置 nginx
-bash "$UPDATE_NGINX_SCRIPT" $1
-
-# 导出合约地址
-kurtosis service exec $1 contracts-1 "cat /opt/zkevm/combined.json |jq {polygonZkEVMBridgeAddress:.polygonZkEVMBridgeAddress, polygonZkEVML2BridgeAddress:.polygonZkEVML2BridgeAddress}" >$CONTRACTS_FILE
+if [ "$DRYRUN" == "true" ]; then
+  echo "[dry-run] envsubst <params.template.yml >$TEMP_CONFIG"
+  echo "[dry-run] kurtosis run --cli-log-level debug -v EXECUTABLE --enclave $1 --args-file $TEMP_CONFIG github.com/Pana/kurtosis-cdk@aa5f6f39dd8fa6157abe5736d81a2c9eda1536fc 2>&1 >$LOG_FILE"
+  echo "[dry-run] set nginx for $1"
+  echo "[dry-run] exported contracts to: $CONTRACTS_FILE"
+  echo "{polygonZkEVMBridgeAddress: \"0x0000000000000000000000000000000000000000\", polygonZkEVML2BridgeAddress: \"0x0000000000000000000000000000000000000000\"}" > $CONTRACTS_FILE
+else
+  envsubst <params.template.yml >$TEMP_CONFIG
+  echo "generated params file: $TEMP_CONFIG"
+  # kurtosis run --cli-log-level debug -v EXECUTABLE --enclave op-eth github.com/wangdayong228/optimism-package@8d97b22f5bce73106fea4d3cc063486cca359928 --args-file "$TEMP_CONFIG" 2>&1 > "$LOG_FILE"
+  kurtosis run --cli-log-level debug -v EXECUTABLE --enclave $1 --args-file $TEMP_CONFIG github.com/Pana/kurtosis-cdk@aa5f6f39dd8fa6157abe5736d81a2c9eda1536fc 2>&1 >$LOG_FILE
+  echo "deployed kurtosis enclave: $1"
+  # 设置 nginx
+  bash "$UPDATE_NGINX_SCRIPT" $1
+  echo "set nginx for $1"
+  # 导出合约地址
+  kurtosis service exec $1 contracts-1 "cat /opt/zkevm/combined.json |jq {polygonZkEVMBridgeAddress:.polygonZkEVMBridgeAddress, polygonZkEVML2BridgeAddress:.polygonZkEVML2BridgeAddress}" >$CONTRACTS_FILE
+  echo "exported contracts to: $CONTRACTS_FILE"
+fi
 
 # 给一直发交易的地址转账
 # cast send --legacy --rpc-url $zkc_l2_rpc --private-key $zkc_l2_pk --value 1000ether 0x8943545177806ED17B9F23F0a21ee5948eCaa776
